shifted react app to frontend folder
