+++ frontend/src/blogs.js
@@ -1,0 +1,137 @@
+const data = [
+  {
+    id: "80C",
+    name: "Section 80C",
+    subtitle: "Deduction for specified investments",
+    summary:
+      "Provides a deduction of up to Rs. 1.5 lakh for investments made in certain specified instruments such as Public Provident Fund (PPF), Equity Linked Savings Scheme (ELSS), National Pension System (NPS), and tax-saving fixed deposits.",
+    detail: [
+      "Section 80C is a provision under the Income Tax Act of India that allows taxpayers to claim deductions on their taxable income for certain types of investments and expenses.",
+      "Under this section, taxpayers can claim a deduction of up to Rs. 1.5 lakh from their taxable income for various investments and expenses, such as:",
+      "1. Investment in tax-saving mutual funds (ELSS)",
+      "2. Investment in Public Provident Fund (PPF)",
+      "3. Investment in National Savings Certificate (NSC)",
+      "4. Investment in Unit Linked Insurance Plan (ULIP)",
+      "5. Payment of life insurance premium Payment of tuition fees for children's education ",
+      "6. Repayment of principal amount of home loan  ",
+      "The benefit of claiming deductions under Section 80C is that it reduces the taxable income of the taxpayer, which in turn reduces their tax liability. ",
+      "This can be particularly beneficial for individuals who are looking to save on their taxes and maximize their savings.    It is important to note that the total deduction claimed under Section 80C cannot exceed Rs. 1.5 lakh in a financial year. ",
+      "Additionally, not all investments and expenses qualify for deduction under this section, so it is important to consult with a tax expert or do thorough research before making any investments or claiming deductions."
+    ]
+  },
+  {
+    id: "80D",
+
+    name: "Section 80D",
+    subtitle: "Deduction for health insurance premiums",
+    summary:
+      "Provides a deduction for the payment of health insurance premiums for self, spouse, children, and parents. The maximum deduction allowed is Rs. 25,000 per annum, which can be increased to Rs. 50,000 if the taxpayer or his/her spouse is a senior citizen.",
+    detail: [
+      "Section 80D is a provision in the Income Tax Act that allows individuals to claim a deduction for the amount they have paid towards medical insurance premium or preventive health check-ups.",
+      "The deduction is available to both individuals and Hindu Undivided Families (HUFs).",
+      "In simpler terms, if you have taken a medical insurance policy for yourself or your family, or undergone a preventive health check-up, you can claim a tax deduction for the amount you have spent on these things, subject to certain limits.",
+      "The amount of deduction you can claim depends on various factors such as your age, the type of insurance policy you have taken, and the amount of premium you have paid.",
+      "The maximum deduction that can be claimed under this section is Rs. 25,000 per year for individuals and Rs. 50,000 per year for senior citizens (individuals aged 60 years or above).",
+      "It is important to note that the deduction is only available for the amount paid towards medical insurance premium or preventive health check-ups and not for any other medical expenses.",
+      "Additionally, the payment should be made through a non-cash mode such as cheque, demand draft, or electronic transfer."
+    ]
+  },
+  {
+    id: "80E",
+
+    name: "Section 80E",
+    subtitle: "Deduction for education loan interest",
+    summary:
+      "Provides a deduction for the payment of interest on an education loan taken for higher studies. There is no limit on the amount of deduction, and it can be claimed for a maximum of eight years.",
+    detail: [
+      "Section 80E of the Income Tax Act allows an individual to claim a deduction on the interest paid on an education loan taken for higher studies for self, spouse, children or for a student for whom the individual is a legal guardian. The deduction is available for a maximum of 8 years or until the interest is fully paid off, whichever is earlier.",
+      "Here are some important points to keep in mind about Section 80E:",
+      "- The deduction is only available for the interest portion of the education loan, and not for the principal amount.",
+      "- The education loan should have been taken from a financial institution or a charitable institution approved by the central government. Loans taken from friends, family or relatives are not eligible for this deduction.",
+      "- The deduction can be claimed only by an individual and not by a Hindu Undivided Family (HUF) or any other entity.",
+      "- The deduction can be claimed only for higher education courses pursued after completion of senior secondary examination or its equivalent from any school, board, or university recognized by the central or state government, including courses pursued outside India.",
+      "- The maximum amount of deduction that can be claimed under Section 80E is the actual amount of interest paid on the education loan during the financial year.",
+      "Overall, Section 80E is a beneficial provision for individuals who have taken an education loan for higher studies as it helps reduce their tax liability and encourages them to pursue further education."
+    ]
+  },
+  {
+    id: "80G",
+
+    name: "Section 80G",
+    subtitle: "Deduction for charitable donations",
+    summary:
+      "Provides a deduction for donations made to certain specified funds and charitable organizations. The amount of deduction allowed varies depending on the nature of the donation and the organization to which it is made.",
+    detail: [
+      "Section 80G of the Income Tax Act provides a deduction to taxpayers for donations made to certain specified funds and charitable organizations. The aim of this provision is to encourage people to donate to social and charitable causes and to support the activities of such organizations.",
+      "The deduction allowed under Section 80G varies depending on the nature of the donation and the organization to which it is made. Some organizations are eligible for a 100% deduction, while others may only be eligible for a 50% deduction. Additionally, there are certain funds and organizations that are not eligible for any deduction.",
+      "To claim the deduction under Section 80G, the taxpayer needs to obtain a receipt or certificate from the organization to which the donation was made. The receipt should contain details such as the name and address of the organization, the amount donated, and the registration number of the organization.",
+      "It is important to note that not all donations are eligible for a deduction under Section 80G. For example, donations made in the form of cash exceeding Rs. 10,000 are not eligible for a deduction, and donations made to political parties are also not eligible for a deduction.",
+      "Overall, Section 80G is a useful provision for taxpayers who want to contribute to charitable causes and also want to reduce their tax liability. By donating to eligible organizations and funds, taxpayers can claim a deduction on their income tax return and support the activities of these organizations at the same time."
+    ]
+  },
+  {
+    id: "80TTA",
+
+    name: "Section 80TTA",
+    subtitle: "Deduction for savings bank interest",
+    summary:
+      "Provides a deduction of up to Rs. 10,000 on the interest earned on savings bank accounts.",
+    detail: [
+      "Section 80TTA is a provision under the Income Tax Act, which allows individuals to claim a deduction of up to Rs. 10,000 on the interest earned on savings bank accounts. This deduction is available to individuals and Hindu Undivided Families (HUFs).",
+      "The deduction is available on the interest earned from savings accounts with banks, co-operative societies, and post offices. However, it does not apply to interest earned on fixed deposits or recurring deposits.",
+      "It is important to note that the deduction is available only to resident individuals and HUFs. Non-resident individuals are not eligible for this deduction.",
+      'To claim the deduction under Section 80TTA, individuals need to disclose the interest earned on their savings bank accounts in the "income from other sources" section of their income tax return. The maximum deduction that can be claimed under this section is Rs. 10,000. If the interest earned on savings bank accounts is less than Rs. 10,000, then the entire interest amount can be claimed as a deduction.',
+      "It is also important to note that the deduction under Section 80TTA is not available to senior citizens, as they are already eligible for a higher deduction of up to Rs. 50,000 on interest income earned from deposits with banks and post offices under Section 80TTB."
+    ]
+  },
+  {
+    id: "80U",
+
+    name: "Section 80U",
+    subtitle: "Deduction for persons with disabilities",
+    summary:
+      "Provides a deduction for persons with disabilities. The deduction amount varies depending on the extent of the disability, and can go up to Rs. 1.25 lakh.",
+    detail: [
+      "Section 80U of the Income Tax Act provides a deduction to an individual who is a resident of India and has a disability. The disability may be physical or mental, and it must be certified by a medical authority.",
+      "The deduction amount varies depending on the extent of the disability and is allowed to an individual if he/she incurs any expenditure for the medical treatment, training or rehabilitation of the person with a disability. The amount of deduction available under this section is as follows:",
+      "If the disability is 40% or more but less than 80%, the individual can claim a deduction of Rs. 75,000.",
+      "If the disability is 80% or more, the individual can claim a deduction of Rs. 1.25 lakh.",
+      "It is important to note that to claim the deduction under this section, the individual must obtain a certificate of disability from a medical authority. The certificate should contain the details of the disability, the extent of disability, and the period for which the disability is likely to continue.",
+      "Additionally, the deduction can only be claimed by the individual and not by any other person such as a spouse or parent. The deduction can be claimed by the individual while filing the income tax return.",
+      "In summary, Section 80U provides a deduction for individuals with disabilities to help them cope with the additional expenses associated with their disability. The deduction amount varies depending on the extent of the disability and is subject to certain conditions."
+    ]
+  },
+  {
+    id: "calculate-taxable-income",
+    name: "Taxable Income",
+    subtitle: "Steps to Calculate Taxable Income",
+    summary:
+      "As per Indian taxation system an individual can calculate his taxable income and corresponding taxes",
+    detail: [
+      "Determine the Gross Total Income (GTI): The first step is to determine the Gross Total Income (GTI), which is the sum of all income earned during the financial year, including salary, rental income, business income, capital gains, and any other income.",
+      "Claim Deductions: Once the GTI is determined, you can claim deductions under various sections of the Income Tax Act, such as Section 80C, 80D, 80G, etc. These deductions are subtracted from the GTI to arrive at the Total Income.",
+      "Calculate Total Income: The Total Income is the income after deducting all eligible deductions from the GTI.",
+      "Calculate Taxable Income: The next step is to calculate the taxable income by subtracting the exemptions and deductions from the Total Income.",
+      "Apply Tax Slab Rates: Once the taxable income is calculated, it is subjected to tax as per the applicable tax slab rates. For the financial year 2022-23, the tax slab rates are as follows:",
+      "Up to Rs. 2.5 lakhs: No tax",
+      "Rs. 2.5 lakhs to Rs. 5 lakhs: 5%",
+      "Rs. 5 lakhs to Rs. 7.5 lakhs: 10%",
+      "Rs. 7.5 lakhs to Rs. 10 lakhs: 15%",
+      "Rs. 10 lakhs to Rs. 12.5 lakhs: 20% ",
+      " Rs. 12.5 lakhs to Rs. 15 lakhs: 25% ",
+      "Above Rs. 15 lakhs: 30% ",
+      "Apply Surcharge and Health and Education Cess: If the taxable income exceeds certain limits, surcharge and health and education cess are applicable as per the Income Tax Act.",
+      "Calculate Tax Liability: Finally, the tax liability is calculated by adding the tax and surcharge, if applicable, and then adding health and education cess to it."
+    ]
+  },
+  {
+    id:"80ccd",
+    name: "Section 80CCD",
+    subtitle: "Deduction in respect of contribution to pension scheme of Central Government",
+
+summary: "Section 80CCD is a provision of the Indian Income Tax Act that allows taxpayers to claim deductions on contributions made towards the National Pension System (NPS) and other pension schemes.",
+
+detail:[ "Section 80CCD has two sub-sections:","Section 80CCD(1):"," This sub-section allows taxpayers to claim deductions on contributions made towards the NPS and other pension schemes."," The deduction limit for section 80CCD(1) is 10% of salary (for salaried individuals) or 10% of gross income (for self-employed individuals) and the maximum deduction limit is Rs. 1.5 lakh."," This deduction is also available to employees who are contributing to the government pension scheme, i.e., the Atal Pension Yojana (APY).","Section 80CCD(1B): This sub-section allows an additional deduction of up to Rs. 50,000 to be claimed on contributions made towards the NPS."," This deduction is over and above the limit allowed under section 80CCD(1)."," However, this deduction is only available to individuals and not to HUFs (Hindu Undivided Families).","In summary, Section 80CCD provides taxpayers with a means of reducing their tax liability by claiming deductions on contributions made towards pension schemes."," The NPS is one such scheme, and the deductions can be claimed under both sub-sections of Section 80CCD"] }
+];
+
+export default data;
