+++ frontend/src/components/Header.jsx
@@ -1,0 +1,22 @@
+import Container from "react-bootstrap/Container";
+import Nav from "react-bootstrap/Nav";
+import Navbar from "react-bootstrap/Navbar";
+
+function ColorSchemesExample() {
+  return (
+    <>
+      <Navbar bg="dark" variant="dark" sticky="top">
+        <Container>
+          <Navbar.Brand href="/">Simpl Tax</Navbar.Brand>
+          <Nav className="me-auto">
+            <Nav.Link href="/calculator">Calculator</Nav.Link>
+            <Nav.Link href="/blogs">Blogs</Nav.Link>
+            <Nav.Link href="/login">Login</Nav.Link>
+          </Nav>
+        </Container>
+      </Navbar>
+    </>
+  );
+}
+
+export default ColorSchemesExample;
